requirements.txt
fastapi == 0.110.3
openai == 0.28.1
supabase == 2.4.6
python-dotenv==1.0.1

